NEW: Added code to allow API call to replace original resource file and optionally bypass versioning [q10932]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@14042 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ pages/upload_plupload.php
@@ -27,6 +27,7 @@ $offset                                 = getvalescaped('offset', '', true);
 $order_by                               = getvalescaped('order_by', '');
 $no_exif_raw                            = getval('no_exif', $metadata_read_default ? '' : 'yes');
 $no_exif                                = $no_exif_raw == "yes" || $no_exif_raw =="1" ? true : false;
+$autorotate                             = getval('autorotate','') != '';
 // This is the archive state for searching, NOT the archive state to be set from the form POST
 $archive                                = getvalescaped('archive', '', true);
 
@@ -523,7 +524,7 @@ if ($_FILES)
                     }
                 else
                     {
-		    debug("PLUPLOAD ERROR- failed  to find temp file " . $_FILES['file']['tmp_name'] . " file received from user " . $username . ",  filename " . $plfilename . ", chunk " . ($chunk+1)  . " of " . $chunks);
+		            debug("PLUPLOAD ERROR- failed  to find temp file " . $_FILES['file']['tmp_name'] . " file received from user " . $username . ",  filename " . $plfilename . ", chunk " . ($chunk+1)  . " of " . $chunks);
                     die('{"jsonrpc" : "2.0", "error" : {"code": 103, "message": "Failed to move uploaded file."}, "id" : "id"}');
                     }
 	} else {
@@ -717,7 +718,7 @@ if ($_FILES)
                             # Log this			
                             daily_stat("Resource upload",$ref);
                             
-                            $status=upload_file($ref,($no_exif=="yes" && getval("exif_override","")==""),false,(getval('autorotate','')!=''),$plupload_upload_location);
+                            $status=upload_file($ref,($no_exif=="yes" && getval("exif_override","")==""),false,$autorotate,$plupload_upload_location);
 
                             if($status && $auto_generated_resource_title_format != '' && !$upload_then_edit)
                                 {
@@ -778,36 +779,26 @@ if ($_FILES)
                         else if ($replace=="" && $replace_resource!="")
                             {
                             // Replacing an existing resource file
-                            daily_stat('Resource upload', $replace_resource);
-
-                            // save original file as an alternative file
-                            if($replace_resource_preserve_option && '' != getval('keep_original', ''))
-								{
-                                save_original_file_as_alternative($replace_resource);    
-                                }  
-								
-                            $status = upload_file($replace_resource, (('yes' == $no_exif) && '' == getval('exif_override', '')), false, ('' != getval('autorotate','')), $plupload_upload_location);
-
-                            hook("additional_replace_existing");
-                            
+                            // Extract data unless user has selected not to extract exif data and there are no per field options set
+                            $no_exif = ('yes' == $no_exif) && '' == getval('exif_override', '');
+                            $keep_original = getval('keep_original', '') != '';
+                            $success = replace_resource_file($replace_resource,$plupload_upload_location,$autorotate,$no_exif,$keep_original);
+                            if (!$success)
+                                {
+                                die('{"jsonrpc" : "2.0", "error" : {"code": 109, "message": "Failed to replace resource file"}, "id" : "' . htmlspecialchars($replace_resource) . '"}');
+                                }
                             if(file_exists($plupload_processed_filepath))
                                 {
                                 unlink($plupload_processed_filepath);
                                 }
 
-                            
-											
-							// Check to see if we need to notify users of this change							
-							if($notify_on_resource_change_days!=0)
-								{								
-								// we don't need to wait for this..
-								ob_flush();flush();	
-								notify_resource_change($replace_resource);
-								}							
                             die('{"jsonrpc" : "2.0", "message" : "' . $lang["replacefile"] . '", "id" : "' . htmlspecialchars($replace_resource) . '"}');
                             }
                     else
                             {
+                            $no_exif = ('yes' == $no_exif) && '' == getval('exif_override', '');
+                            $keep_original = getval('keep_original', '') != '';
+                               
                             if (!isset($batch_replace_col) || $batch_replace_col == 0)
                                 {
                                 $conditions = array();
@@ -829,28 +820,25 @@ if ($_FILES)
                                 $target_resource=array_values(array_intersect($target_resource,$replace_resources));
                                 if(count($target_resource)==1  && !resource_file_readonly($target_resource[0]))
 									{
-									// A single resource has been found with the same filename
-									daily_stat("Resource upload",$target_resource[0]);
-									$status=upload_file($target_resource[0],($no_exif=="yes" && getval("exif_override","")==""),false,(getval('autorotate','')!=''), $plupload_upload_location); # Upload to the specified ref.
-									if(file_exists($plupload_processed_filepath))
+                                    // A single resource has been found with the same filename                                    
+                                    $success = replace_resource_file($target_resource[0],$plupload_upload_location,$no_exif,$autorotate,$keep_original);
+                                    if (!$success)
+                                        {
+                                        die('{"jsonrpc" : "2.0", "error" : {"code": 109, "message": "Failed to replace resource file"}, "id" : "' . htmlspecialchars($target_resource[0]) . '"}');
+                                        }
+                                    unlink($plupload_upload_location);
+                                    if(file_exists($plupload_processed_filepath))
                                         {
                                         unlink($plupload_processed_filepath);
                                         }
-                                    // Check to see if we need to notify users of this change							
-									if($notify_on_resource_change_days!=0)
-										{								
-										// we don't need to wait for this..
-										ob_flush();flush();
-										
-										notify_resource_change($target_resource[0]);
-										}
+
 									die('{"jsonrpc" : "2.0", "message" : "' . $lang["upload_success"] . ' - ' . $lang["replacefile"] . '", "id" : "' . htmlspecialchars($target_resource[0]) . '"}');
 									}
 								elseif(count($target_resource)==0)
 									{
 									// No resource found with the same filename
 									header('Content-Type: application/json');
-                                    unlink($plfilepath);
+                                    unlink($plupload_upload_location);
                                     if(file_exists($plupload_processed_filepath))
                                         {
                                         unlink($plupload_processed_filepath);
@@ -866,7 +854,18 @@ if ($_FILES)
 										{
 										foreach ($target_resource as $replaced)
 											{
-											$status = upload_file($replaced, ('yes' == $no_exif && '' == getval('exif_override', '')), false, ('' != getval('autorotate', '')), $plupload_upload_location);
+                                            $success = replace_resource_file($replaced,$plupload_upload_location,$no_exif,$autorotate,$keep_original);
+                                            if (!$success)
+                                                {
+                                                die('{"jsonrpc" : "2.0", "error" : {"code": 109, "message": "Failed to replace resource file"}, "id" : "' . htmlspecialchars($replaced) . '"}');
+                                                }
+                                            
+                                            unlink($plupload_upload_location);
+                                            if(file_exists($plupload_processed_filepath))
+                                                {
+                                                unlink($plupload_processed_filepath);
+                                                }
+											$status = upload_file($replaced, ('yes' == $no_exif && '' == getval('exif_override', '')), false, $autorotate, $plupload_upload_location);
 											}
                                         unlink($plfilepath);
                                         die('{"jsonrpc" : "2.0", "message" : "' . $lang["replacefile"] . '", "id" : "' . $resourcelist . '"}');
@@ -875,7 +874,7 @@ if ($_FILES)
 										{
 										// Multiple resources found with the same filename
 										header('Content-Type: application/json');
-                                        unlink($plfilepath);
+                                        unlink($plupload_upload_location);
                                         if(file_exists($plupload_processed_filepath))
                                             {
                                             unlink($plupload_processed_filepath);
@@ -906,12 +905,12 @@ if ($_FILES)
                                             # Save the original file as an alternative file?                                            
                                             $keep_original = getval('keep_original', '');
                                             $save_original = ($keep_original == 1) ? save_original_file_as_alternative($ref) : true;
-                                            if (!$save_original)
+                                            
+                                            $success = replace_resource_file($ref,$plupload_upload_location,$no_exif,$autorotate,$keep_original);
+                                            if (!$success)
                                                 {
-                                                die('{"jsonrpc" : "2.0", "error" : {"code": 109, "message": "' . $lang["error_saveorigalternative"] . '"}, "id" : "' . htmlspecialchars($ref) . '"}');
+                                                die('{"jsonrpc" : "2.0", "error" : {"code": 109, "message": "Failed to replace resource file"}, "id" : "' . htmlspecialchars($ref) . '"}');
                                                 }
-
-                                            $status = upload_file($ref, ('yes' == $no_exif && '' == getval('exif_override', '')), false, ('' != getval('autorotate', '')), $plupload_upload_location);
 
                                             die('{"jsonrpc" : "2.0", "message" : "' . $lang["replacefile"] . '", "id" : "' . htmlspecialchars($ref) . '"}');
                                             }
@@ -1211,7 +1210,7 @@ var pluploadconfig = {
                                   uploader.bind('UploadComplete', function(up, files) {
                                   if(!upRedirBlock)
                                       {
-                                      CentralSpaceLoad('<?php echo $redirecturl ?>',true);
+                                      //CentralSpaceLoad('<?php echo $redirecturl ?>',true);
                                       }
                                   upRedirBlock = false; 
                                   });
